Revert "Fix rendering of template fields with start from trigger"
This reverts commit 4362c25e5cb96e3b6c99497b682d4dfe23a86aba from https://github.com/apache/airflow/pull/53071.

The original change violated Airflow's architectural separation by making
the Triggerer supervisor process load DAGs and attempt template rendering.
This breaks with mapped operators that use lazy sequences requiring
SUPERVISOR_COMMS, which only exists in task execution context.

The Triggerer should operate purely on trigger database records and
pre-computed trigger kwargs, not attempt runtime DAG operations.

Fixes triggerer failures with mapped operators using XCom dependencies.
