+++ airflow-core/src/airflow/jobs/triggerer_job_runner.py
@@ -43,7 +43,6 @@ from airflow.configuration import conf
 from airflow.executors import workloads
 from airflow.jobs.base_job_runner import BaseJobRunner
 from airflow.jobs.job import perform_heartbeat
-from airflow.models import DagBag
 from airflow.models.trigger import Trigger
 from airflow.sdk.api.datamodels._generated import HITLDetailResponse
 from airflow.sdk.execution_time.comms import (
@@ -606,45 +605,6 @@ class TriggerRunnerSupervisor(WatchedSubprocess):
         trigger set.
         """
         render_log_fname = log_filename_template_renderer()
-        dag_bag = DagBag(collect_dags=False)
-
-        def expand_start_trigger_args(trigger: Trigger) -> Trigger:
-            task = dag_bag.get_dag(trigger.task_instance.dag_id).get_task(trigger.task_instance.task_id)
-            if task.template_fields:
-                trigger.task_instance.refresh_from_task(task)
-                context = trigger.task_instance.get_template_context()
-                task.render_template_fields(context=context)
-                start_trigger_args = task.expand_start_trigger_args(context=context)
-                if start_trigger_args:
-                    trigger.kwargs = start_trigger_args.trigger_kwargs
-            return trigger
-
-        def create_workload(trigger: Trigger) -> workloads.RunTrigger:
-            if trigger.task_instance:
-                log_path = render_log_fname(ti=trigger.task_instance)
-
-                trigger = expand_start_trigger_args(trigger)
-
-                ser_ti = workloads.TaskInstance.model_validate(trigger.task_instance, from_attributes=True)
-                # When producing logs from TIs, include the job id producing the logs to disambiguate it.
-                self.logger_cache[new_id] = TriggerLoggingFactory(
-                    log_path=f"{log_path}.trigger.{self.job.id}.log",
-                    ti=ser_ti,  # type: ignore
-                )
-
-                return workloads.RunTrigger(
-                    classpath=trigger.classpath,
-                    id=new_id,
-                    encrypted_kwargs=trigger.encrypted_kwargs,
-                    ti=ser_ti,
-                    timeout_after=trigger.task_instance.trigger_timeout,
-                )
-            return workloads.RunTrigger(
-                classpath=trigger.classpath,
-                id=new_id,
-                encrypted_kwargs=trigger.encrypted_kwargs,
-                ti=None,
-            )
 
         known_trigger_ids = (
             self.running_triggers.union(x[0] for x in self.events)
@@ -682,7 +642,26 @@ class TriggerRunnerSupervisor(WatchedSubprocess):
                 )
                 continue
 
-            workload = create_workload(new_trigger_orm)
+            workload = workloads.RunTrigger(
+                classpath=new_trigger_orm.classpath,
+                id=new_id,
+                encrypted_kwargs=new_trigger_orm.encrypted_kwargs,
+                ti=None,
+            )
+            if new_trigger_orm.task_instance:
+                log_path = render_log_fname(ti=new_trigger_orm.task_instance)
+
+                ser_ti = workloads.TaskInstance.model_validate(
+                    new_trigger_orm.task_instance, from_attributes=True
+                )
+                # When producing logs from TIs, include the job id producing the logs to disambiguate it.
+                self.logger_cache[new_id] = TriggerLoggingFactory(
+                    log_path=f"{log_path}.trigger.{self.job.id}.log",
+                    ti=ser_ti,  # type: ignore
+                )
+
+                workload.ti = ser_ti
+                workload.timeout_after = new_trigger_orm.task_instance.trigger_timeout
 
             to_create.append(workload)
