+++ airflow-core/src/airflow/triggers/base.py
@@ -19,6 +19,8 @@ from __future__ import annotations
 import abc
 import json
 from collections.abc import AsyncIterator
+from dataclasses import dataclass
+from datetime import timedelta
 from typing import Annotated, Any
 
 import structlog
@@ -31,25 +33,20 @@ from pydantic import (
 )
 
 from airflow.utils.log.logging_mixin import LoggingMixin
-from airflow.utils.module_loading import import_string
 from airflow.utils.state import TaskInstanceState
 
 log = structlog.get_logger(logger_name=__name__)
 
 
-def __getattr__(name: str):
-    if name == "StartTriggerArgs":
-        import warnings
+@dataclass
+class StartTriggerArgs:
+    """Arguments required for start task execution from triggerer."""
 
-        warnings.warn(
-            "airflow.triggers.base.StartTriggerArgs is deprecated. "
-            "Use airflow.sdk.bases.trigger.StartTriggerArgs instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return import_string(f"airflow.sdk.bases.trigger.{name}")
-
-    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
+    trigger_cls: str
+    next_method: str
+    trigger_kwargs: dict[str, Any] | None = None
+    next_kwargs: dict[str, Any] | None = None
+    timeout: timedelta | None = None
 
 
 class BaseTrigger(abc.ABC, LoggingMixin):
