+++ task-sdk/docs/api.rst
@@ -72,8 +72,6 @@ Bases
 
 .. autoapiclass:: airflow.sdk.BaseHook
 
-.. autoapiclass:: airflow.sdk.StartTriggerArgs
-
 Connections & Variables
 -----------------------
 .. autoapiclass:: airflow.sdk.Connection
