+++ task-sdk/src/airflow/sdk/__init__.py
@@ -39,7 +39,6 @@ __all__ = [
     "ObjectStoragePath",
     "Param",
     "PokeReturnValue",
-    "StartTriggerArgs",
     "TaskGroup",
     "TriggerRule",
     "Variable",
@@ -68,7 +67,6 @@ if TYPE_CHECKING:
     from airflow.sdk.bases.operator import BaseOperator, chain, chain_linear, cross_downstream
     from airflow.sdk.bases.operatorlink import BaseOperatorLink
     from airflow.sdk.bases.sensor import BaseSensorOperator, PokeReturnValue
-    from airflow.sdk.bases.trigger import StartTriggerArgs
     from airflow.sdk.definitions.asset import Asset, AssetAlias, AssetAll, AssetAny, AssetWatcher
     from airflow.sdk.definitions.asset.decorators import asset
     from airflow.sdk.definitions.asset.metadata import Metadata
@@ -106,7 +104,6 @@ __lazy_imports: dict[str, str] = {
     "Param": ".definitions.param",
     "PokeReturnValue": ".bases.sensor",
     "SecretCache": ".execution_time.cache",
-    "StartTriggerArgs": ".bases.trigger",
     "TaskGroup": ".definitions.taskgroup",
     "TriggerRule": ".api.datamodels._generated",
     "Variable": ".definitions.variable",
