+++ task-sdk/src/airflow/sdk/__init__.pyi
@@ -29,7 +29,6 @@ from airflow.sdk.bases.sensor import (
     BaseSensorOperator as BaseSensorOperator,
     PokeReturnValue as PokeReturnValue,
 )
-from airflow.sdk.bases.trigger import StartTriggerArgs as StartTriggerArgs
 from airflow.sdk.definitions.asset import (
     Asset as Asset,
     AssetAlias as AssetAlias,
@@ -79,7 +78,6 @@ __all__ = [
     "Param",
     "PokeReturnValue",
     "SecretCache",
-    "StartTriggerArgs",
     "TaskGroup",
     "TriggerRule",
     "Variable",
